server/requirements.txt
acme==3.0.1
blinker==1.8.2
certbot==3.0.1
certbot-nginx==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
ConfigArgParse==1.7
configobj==5.0.9
cryptography==44.0.0
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
josepy==1.14.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
parsedatetime==2.6
pycparser==2.22
pyOpenSSL==24.3.0
pyparsing==3.2.1
pyRFC3339==2.0.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
urllib3==2.3.0
Werkzeug==3.0.6
